Fix compilation on macOS
On macOS, the Python extension needs to be built as a dynamic library instead of
a bundle (which is the default produced by setuptools). The README contains
instructions for this now. Furthermore, it's necessary to edit a few of the
library header fields after copying the file; this is done with an extra custom
command per library.

Also, don't link against `_dl*.so` anymore.